requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.28.15
botocore==1.31.15
certifi==2023.7.22
charset-normalizer==3.2.0
colour==0.1.5
cryptography==41.0.7
cython==3.0.0
django-appconf==1.0.5
django-compressor==4.4
Django==4.0
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
idna==3.4
jmespath==1.0.1
numpy==1.24.4
pandas==1.3.5
psycopg2==2.9.3
pyjwt==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
